Теперь можно передать порт в качестве параметра, если параметр не передан, то запускаемся на стандартном порту(статическая переменная DEFAULT_SERVER_PORT).

+++ src/main/tests/WebPageTest.java
@@ -73,7 +73,7 @@ public class WebPageTest {
         Map<String, Integer> checkArr = checkPage(result);
         // Печать результата
         for (String current : checkArr.keySet()) {
-            Assert.assertNotSame(ErrText + current, checkArr.get(current).intValue(), -1);
+            Assert.assertNotSame(ErrText + current, checkArr.get(current), -1);
         }
     }
 
@@ -92,7 +92,7 @@ public class WebPageTest {
         Map<String, Integer> checkArr = checkPage(result);
         // Вывод
         for (String current : checkArr.keySet()) {
-            Assert.assertNotSame(ErrText + current, checkArr.get(current).intValue(), -1);
+            Assert.assertNotSame(ErrText + current, checkArr.get(current), -1);
         }
     }
